visibility structure
