+++ src/dashboard/dashboard/TruckVisibiltyReport.js
@@ -5,6 +5,41 @@ import TruckStatus from "./TruckStatus";
 import Map from "../../images/Map.png";
 
 function TruckVisibiltyReport() {
+  const trucks = [
+    {
+      id: "1",
+      name: "AXL 001",
+      location: {
+        name: "Ibafo Deposit",
+        kms: 21305,
+      },
+      date: "2023-04-07T12:30:00Z",
+      operationalStatus: "Running",
+      datStatus: "Active",
+    },
+    {
+      id: "2",
+      name: "AXL 002",
+      location: {
+        name: "Ota Deposit",
+        kms: 32040,
+      },
+      date: "2023-04-07T14:00:00Z",
+      operationalStatus: "Stopped",
+      datStatus: "Inactive",
+    },
+    {
+      id: "3",
+      name: "AXL 003",
+      location: {
+        name: "Lekki Deposit",
+        kms: 10450,
+      },
+      date: "2023-04-07T09:45:00Z",
+      operationalStatus: "Running",
+      datStatus: "Active",
+    },
+  ];
   return (
     <div className="md:flex gap-5 py-8">
       <div className="space-y-3 basis-1/2">
@@ -28,10 +63,10 @@ function TruckVisibiltyReport() {
             <p className="text-center">Available</p>
             <p className="text-center">Unavailable</p>
           </div>
-          <TruckStatus />
-          <TruckStatus />
-          <TruckStatus />
-          <TruckStatus />
+
+          {trucks.map((truck, index) => (
+            <TruckStatus key={index} truck={truck} />
+          ))}
         </div>
         <button className="btn btn-primary float-right">Donwload Report</button>
       </div>
