+++ src/dashboard/dashboard/truckContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const TruckContext = createContext(null);
+
+export default TruckContext;
