+++ src/dashboard/operations/shortageResolution/ProductDetails.jsx
@@ -1,60 +1,10 @@
-import React, { useState } from "react";
-import LiquidHeight from "./LiquidHeight";
-import Meter from "./Meter";
-import UllageParameter from "./UllageParameter";
+import TruckResolution from "./TruckResolution";
+import { useParams } from "react-router-dom";
+import { useState, useEffect } from "react";
+import endpoint from "../../../utils/endpoints/endpoint";
+import useFetch from "../../../useFetch";
 
-function ProductDetails({ trucks, formData, setFormData }) {
-  const handleCalibrationTypeChange = (event) => {
-    const newCalibrationType = parseInt(event.target.value);
-    setFormData((prevFormData) => ({
-      ...prevFormData,
-      resolutions: [
-        {
-          ...prevFormData.resolutions[0],
-          calibration: {
-            ...prevFormData.resolutions[0].calibration,
-            calibrationType: newCalibrationType,
-          },
-        },
-      ],
-    }));
-  };
-  const options = [
-    { value: 0, parameterId: 1, parameter: "Ullage" },
-    { value: 1, parameterId: 2, parameter: "Liquid Height" },
-    { value: 2, parameterId: 3, parameter: "Meter" },
-  ];
-
-  const calibrationType = () => {
-    if (formData.resolutions[0].calibration.calibrationType === 0) {
-      return <UllageParameter formData={formData} setFormData={setFormData} />;
-    } else if (formData.resolutions[0].calibration.calibrationType === 1) {
-      return <LiquidHeight formData={formData} setFormData={setFormData} />;
-    } else if (formData.resolutions[0].calibration.calibrationType === 2) {
-      return <Meter formData={formData} setFormData={setFormData} />;
-    } else {
-      return "";
-    }
-  };
-  const [disableFields, setDisableFields] = useState({
-    company: false,
-    do: false,
-    customer: false,
-  });
-
-  const handleResponsiblePartyChange = (e) => {
-    const value = parseInt(e.target.value);
-    setFormData({
-      ...formData,
-      resolutions: [
-        {
-          ...formData.resolutions[0],
-          source: value,
-        },
-      ],
-    });
-  };
-
+function ProductDetails({ trucks, setFormData, formData }) {
   return (
     <div>
       <form className="grid gap-5 text-primary my-5">
@@ -194,168 +144,6 @@ function ProductDetails({ trucks, formData, setFormData }) {
               name=""
               id=""
             />
-          </label>
-        </fieldset>
-        <h4 className="text-xl text-primary font-medium">
-          ACL Shortage Resolution
-        </h4>
-        <fieldset className="grid gap-3 md:grid-cols-2 lg:grid-cols-3 items-end">
-          <label htmlFor="">
-            Resolution Parameter
-            <select
-              className="w-full select select-primary"
-              name=""
-              id=""
-              value={formData.resolutions[0].calibration.calibrationType}
-              onChange={handleCalibrationTypeChange}
-            >
-              <option>select Parament</option>
-              {options.map((opt) => (
-                <option key={opt.parameterId} value={opt.value}>
-                  {opt.parameter}
-                </option>
-              ))}
-            </select>
-          </label>
-
-          <label htmlFor="">
-            Shortage Reason
-            <select
-              className="w-full select select-primary"
-              value={formData.resolutions[0].reason}
-              name=""
-              id=""
-              onChange={(e) =>
-                setFormData({
-                  ...formData,
-                  resolutions: [
-                    {
-                      ...formData.resolutions[0],
-                      reason: parseInt(e.target.value),
-                    },
-                  ],
-                })
-              }
-            >
-              <option>select Shortage Reason</option>
-              <option value={0}>Calibration Chart</option>
-              <option value={1}> Theft</option>
-              <option value={2}>Ullage Ground</option>
-              <option value={3}>Short Loading</option>
-              <option value={4}> Other</option>
-            </select>
-          </label>
-          <label htmlFor="">
-            Responsible Party
-            <select
-              value={formData.resolutions[0].source}
-              className="w-full select select-primary"
-              name=""
-              id=""
-              onChange={handleResponsiblePartyChange}
-            >
-              <option>select Responsible Party</option>
-              <option value={0}>Company</option>
-              <option value={1}>DO</option>
-              <option value={2}>Customer</option>
-            </select>
-          </label>
-
-          <label htmlFor="">
-            Amount Charged to Responsible Party
-            <input
-              className="input input-primary w-full"
-              type="number"
-              name=""
-              value={formData.resolutions[0].sourceCharge}
-              id=""
-              disabled={disableFields.company}
-              onChange={(e) =>
-                setFormData({
-                  ...formData,
-                  resolutions: [
-                    {
-                      ...formData.resolutions[0],
-                      sourceCharge: parseInt(e.target.value),
-                    },
-                  ],
-                })
-              }
-            />
-          </label>
-        </fieldset>
-        {calibrationType()}
-        <label htmlFor="">
-          Remarks
-          <textarea
-            className="textarea textarea-primary w-full"
-            value={formData.resolutions[0].remarks}
-            name=""
-            id=""
-            type="text"
-            onChange={(e) =>
-              setFormData({
-                ...formData,
-                resolutions: [
-                  {
-                    ...formData.resolutions[0],
-                    remarks: e.target.value,
-                  },
-                ],
-              })
-            }
-            rows="5"
-          ></textarea>
-        </label>
-
-        <fieldset className="grid gap-3 md:grid-cols-2 items-end">
-          <label htmlFor="">
-            Resolution Status
-            <select
-              className="w-full select select-primary"
-              name=""
-              id=""
-              value={formData.resolutions[0].status}
-              onChange={(e) =>
-                setFormData({
-                  ...formData,
-                  resolutions: [
-                    {
-                      ...formData.resolutions[0],
-                      status: parseInt(e.target.value),
-                    },
-                  ],
-                })
-              }
-            >
-              <option value="">Select Resolution Status</option>
-              <option value={0}>Open</option>
-              <option value={1}>Closed</option>
-            </select>
-          </label>
-          <label htmlFor="">
-            Shortage Refund Status
-            <select
-              className="w-full select select-primary"
-              name=""
-              id=""
-              value={formData.resolutions[0].refundStatus}
-              onChange={(e) =>
-                setFormData({
-                  ...formData,
-                  resolutions: [
-                    {
-                      ...formData.resolutions[0],
-                      refundStatus: parseInt(e.target.value),
-                    },
-                  ],
-                })
-              }
-            >
-              <option value="">select Shortage Refund Status</option>
-              <option value={0}>Pending</option>
-              <option value={1}>Completed</option>
-            </select>
           </label>
         </fieldset>
       </form>
