+++ src/data/DashMenuData.js
@@ -23,6 +23,14 @@ export const OPERATIONS_ROUTES = [
     label: " Truck Assignment",
     route: "/dashboard/TruckAssignment",
   },
+  {
+    label: " Truck Visibility Report",
+    route: "/dashboard/TruckVisibilityReport",
+  },
+  {
+    label: " Truck Visibility Structure",
+    route: "/dashboard/TruckDetails",
+  },
 ];
 export const HEALTHSAFETY_ROUTES = [
   {
