+++ src/data/trucksDemoData.js
@@ -1,0 +1,32 @@
+export const trucks = [
+  {
+    name: "AXL 001",
+    location: {
+      name: "Ibafo Deposit",
+      kms: 21305,
+    },
+    date: "2023-04-07T12:30:00Z",
+    operationalStatus: "Running",
+    datStatus: "Active",
+  },
+  {
+    name: "AXL 002",
+    location: {
+      name: "Ota Deposit",
+      kms: 32040,
+    },
+    date: "2023-04-07T14:00:00Z",
+    operationalStatus: "Stopped",
+    datStatus: "Inactive",
+  },
+  {
+    name: "AXL 003",
+    location: {
+      name: "Lekki Deposit",
+      kms: 10450,
+    },
+    date: "2023-04-07T09:45:00Z",
+    operationalStatus: "Running",
+    datStatus: "Active",
+  },
+];
